coap_module/requirements.txt
aiocoap==0.4.7
cbor-diag==1.0.1
cbor2==5.4.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Cython==0.29.36
DTLSSocket==0.1.16
fe25519==1.4.2
filelock==3.12.2
ge25519==1.4.3
idna==3.4
influxdb-client==1.37.0
mysqlclient==2.1.1
pika==1.3.2
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.4
websockets==11.0.3
mysql-connector-python